Prashanth code review comments part 1
