+++ code/messaging/src/main/java/com/adobe/marketing/mobile/aepcomposeui/AepUIConstants.kt
@@ -13,18 +13,17 @@ package com.adobe.marketing.mobile.aepcomposeui
 
 import androidx.compose.ui.text.font.FontWeight
 
-internal class AepUIConstants {
+internal object AepUIConstants {
 
-    internal class SmallImageCard {
-        internal class DefaultStyle {
-            companion object {
-                const val TITLE_TEXT_SIZE = 15
-                val TITLE_FONT_WEIGHT = FontWeight.Medium
-                const val BODY_TEXT_SIZE = 13
-                val BODY_FONT_WEIGHT = FontWeight.Normal
-                const val BUTTON_TEXT_SIZE = 13
-                val BUTTON_FONT_WEIGHT = FontWeight.Normal
-            }
+    internal object SmallImageCard {
+        internal object DefaultStyle {
+            const val TITLE_TEXT_SIZE = 15
+            val TITLE_FONT_WEIGHT = FontWeight.Medium
+            const val BODY_TEXT_SIZE = 13
+            val BODY_FONT_WEIGHT = FontWeight.Normal
+            const val BUTTON_TEXT_SIZE = 13
+            val BUTTON_FONT_WEIGHT = FontWeight.Normal
+            const val SPACING = 8
         }
     }
 }
