+++ code/messaging/src/main/java/com/adobe/marketing/mobile/aepuitemplates/uimodels/AepDismissButton.kt
@@ -19,4 +19,5 @@ package com.adobe.marketing.mobile.aepuitemplates.uimodels
  * - `simple`: A simple dismiss icon.
  * - `circle`: A circular dismiss icon.
  */
+// TODO change this class to be generic to IconButton
 data class AepDismissButton(val style: String? = null)
